requirements.txt
PySide6==6.6.1
sqlalchemy==2.0.29
python-dotenv==1.0.1
pymysql==1.1.0
reportlab==4.0.9
python-dateutil==2.9.0.post0
python-docx==1.1.0
openpyxl==3.1.2
pdfkit==1.0.0
pandas==2.2.2
numpy==1.26.4


